move nav to own file too
